bullets work now, but i think i'll need to scrap the project, and actually _plan_ the thing the next time

+++ source/global.h
@@ -1,6 +1,8 @@
+#include <ncurses.h>
+
 #ifndef __GLOBAL_H__
 	#define __GLOBAL_H__
-	#define CLOCKS_TO_SEC(clocksGivenAsInput) ((clocksGivenAsInput) / CLOCKS_PER_SEC)
+	#define CLOCKS_TO_SEC(clocksGivenAsInput) ((double) (clocksGivenAsInput) / CLOCKS_PER_SEC)
 
 	typedef struct Point {
 		int16_t y;
